pyproject.toml
[project]
name = "my-dou"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "curl-cffi>=0.15.0",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
]
